app/requirements.txt
Django>>4.2
Adafruit_Python_DHT
pyTelegramBotAPI
djangorestframework
